Draft version of JSON output from zimcheck
The DANGLING_LINKS message is problematic since it may include a list of
(non-normalized) links. Since mustache doesn't provide an API for
iterating over its object data (when its encapsulated inside basic_data)
handling the list of links in JSON mode was impossible. A workaround in
this commit is to use a multi-line string instead of a list.